+++ src/main/kotlin/BukkitGradlePlugin.kt
@@ -1,15 +1,17 @@
 package ru.endlesscode.bukkitgradle
 
-import org.gradle.api.JavaVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.plugins.JavaPluginExtension
 import org.gradle.api.tasks.compile.JavaCompile
+import org.gradle.jvm.toolchain.JavaLanguageVersion
 import org.gradle.kotlin.dsl.*
 import ru.endlesscode.bukkitgradle.dependencies.Dependencies
 import ru.endlesscode.bukkitgradle.meta.PluginMetaPlugin
 import ru.endlesscode.bukkitgradle.meta.extension.PluginMetaImpl
+import ru.endlesscode.bukkitgradle.meta.util.MinecraftVersion
 import ru.endlesscode.bukkitgradle.meta.util.StringUtils
+import ru.endlesscode.bukkitgradle.meta.util.parsedApiVersion
 import ru.endlesscode.bukkitgradle.server.DevServerPlugin
 import ru.endlesscode.bukkitgradle.server.extension.ServerConfigurationImpl
 
@@ -28,7 +30,7 @@ public class BukkitGradlePlugin : Plugin<Project> {
 
     /** Adds needed plugins. */
     private fun Project.addPlugins() {
-        extensions.create<BukkitExtension>("bukkit", configurePluginMeta(), ServerConfigurationImpl())
+        val bukkit = extensions.create<BukkitExtension>("bukkit", configurePluginMeta(), ServerConfigurationImpl())
 
         with(plugins) {
             apply("java")
@@ -37,8 +39,9 @@ public class BukkitGradlePlugin : Plugin<Project> {
         }
 
         extensions.configure<JavaPluginExtension> {
-            sourceCompatibility = JavaVersion.VERSION_1_8
-            targetCompatibility = JavaVersion.VERSION_1_8
+            toolchain {
+                languageVersion.convention(provider { resolveRecommendedJavaVersion(bukkit.parsedApiVersion) })
+            }
         }
     }
 
@@ -48,7 +51,7 @@ public class BukkitGradlePlugin : Plugin<Project> {
             description.convention(provider { project.description })
             main.convention(name.map { "${project.group}.${StringUtils.toPascalCase(it)}" })
             version.convention(provider { project.version.toString() })
-            apiVersion.convention(provider { StringUtils.parseApiVersion(bukkit.apiVersion) })
+            apiVersion.convention(provider { resolveDefaultApiVersion(bukkit.parsedApiVersion) })
             url.convention(provider { providers.gradleProperty("url").orNull })
         }
     }
@@ -65,5 +68,21 @@ public class BukkitGradlePlugin : Plugin<Project> {
         repositories {
             mavenCentral()
         }
+    }
+
+    private fun resolveDefaultApiVersion(version: MinecraftVersion): String? = when {
+        // "API version" has been introduced in Spigot 1.13
+        version < MinecraftVersion.V1_13_0 -> null
+        // From 1.20.5 and onward, a patch version is supported.
+        version < MinecraftVersion.V1_20_5 -> version.withoutPatch().toString()
+        else -> version.toString()
+    }
+
+    // See: https://docs.papermc.io/paper/getting-started#requirements
+    private fun resolveRecommendedJavaVersion(version: MinecraftVersion): JavaLanguageVersion = when {
+        version >= MinecraftVersion.V1_17_1 -> JavaLanguageVersion.of(21)
+        version >= MinecraftVersion.V1_16_5 -> JavaLanguageVersion.of(16)
+        version >= MinecraftVersion.V1_12_0 -> JavaLanguageVersion.of(11)
+        else -> JavaLanguageVersion.of(8)
     }
 }